vectorization example through ap_uint

+++ aie/src/my_kernel_1.h
@@ -2,4 +2,4 @@
 #pragma once
 #include <adf.h>
 
-void my_kernel_function (input_stream<float>* restrict input, output_stream<float>* restrict output);
+void my_kernel_function (input_stream<int32_t>* restrict input, output_stream<int32_t>* restrict output);
